requirements.txt
telethon==1.32.1
aiosqlite==0.19.0
python-dotenv==1.0.1
aiohttp==3.9.3
loguru==0.7.2 